Assets Explorer separation
