+++ editor/js/Sidebar.Materials.js
@@ -8,7 +8,7 @@ function SidebarMaterials(editor, materials) {
     function applyMaterial(path) {
         const glbUrl = `${ASSETS_BASE_URL}/${path}${ext}`;
 
-        editor.loader.addMaterial(glbUrl)
+        editor.explorer.addMaterial(glbUrl)
     };
 
     // Layout
